added sorting by product id (#38)
In Demo Shop UI - we are sorting products client side.

This led to weird selectors (nth element) being chosen when creating new products, as the sorting was changing the state of the DOM and selecting the wrong product item.

To fix this - I added support for sorting by product_id to the Demo Shop Products API. I will then now follow this up with a change to Demo Shop UI that removes the test ID from View Details and replaces the client side sorting with API call query sorting parameters.

Tested this locally

Co-authored-by: Sahil Lamba <slamba@Sahils-MacBook-Pro-2.local>